Tested somewhat successfully. Unable to take picture due to full storage, but the app switched to the camera nicely, and back again.
Note that this is almost entirely Android Developers code at this point.

+++ CameraTest/app/src/main/java/wsv759/cameratest/MainActivity.java
@@ -193,10 +193,10 @@ public class MainActivity extends ActionBarActivity {
     /*
      * Return the fraction of free space available for internal storage by this app.
      */
-    public long getFractionFreeSpace()
+    public double getFractionFreeSpace()
     {
         File dataDir = Environment.getDataDirectory();
-        return dataDir.getFreeSpace() / dataDir.getTotalSpace();
+        return (double) dataDir.getFreeSpace() / dataDir.getTotalSpace();
     }
 
     /*
